requirement.txt
google-auth
djangorestframework
python-dotenv
djongo
requests
django-cors-headers
bcrypt
Django
pymongo
certifi
